requirements.txt
Django==1.6
South==0.8.2
psycopg2==2.5.1
django-tastypie==0.10.0
mimeparse==0.1.3
django-annoying==0.7.7
model_mommy==1.1
django_compressor
